added eslint plugin for react hooks. Separated block transactions into a separate component to simplify logic. Need to work on not losing focus on input when typing in block transactions
