+++ src/components/Block/BlockChain.tsx
@@ -1,6 +1,8 @@
 import React, { useContext } from "react";
 
 import Block from "./Block";
+import BlockTrans from "./BlockTrans";
+
 import { AppContext } from "../../context/AppContext";
 import { IAction, IState } from "../../typings/AppTypes";
 
@@ -11,8 +13,11 @@ export default function BlockChain(): JSX.Element {
 
   return (
     <div className="mx-3 row flex-nowrap overflow-auto">
-      {state.chain.map((block) => (
-        <Block block={block} key={Math.random()} />
+      {state.chain.map((block, i) => (
+        <div className="block" key={i}>
+          <Block block={block} />
+          {block.showTrans && <BlockTrans block={block} />}
+        </div>
       ))}
     </div>
   );
